+++ lib/drawer_scaffold.dart
@@ -30,12 +30,10 @@ class DrawerScaffold extends StatefulWidget {
   final Widget bottomSheet;
   final bool extendBodyBehindAppBar;
   final List<Widget> persistentFooterButtons;
-
   final bool primary;
-
   final bool resizeToAvoidBottomInset;
-
   final bool resizeToAvoidBottomPadding;
+
   DrawerScaffold({
     this.appBar,
     this.contentShadow = const [
@@ -47,7 +45,7 @@ class DrawerScaffold extends StatefulWidget {
       ),
     ],
     this.drawers,
-    this.cornerRadius = 10.0,
+    this.cornerRadius = 16.0,
     this.contentView,
     this.controller,
     this.extendedBody = false,
@@ -113,14 +111,25 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
     super.didUpdateWidget(oldWidget);
   }
 
+  MenuController createController(SideDrawer d) {
+    return MenuController(
+      d.direction,
+      vsync: this,
+    )..addListener(() => setState(() {}));
+  }
+
   assignContoller() {
-    
-    menuControllers = widget.drawers
-        .map((d) => MenuController(
-              d.direction,
-              vsync: this,
-            )..addListener(() => setState(() {})))
-        .toList();
+    if (menuControllers == null)
+      menuControllers ??= widget.drawers.map(createController).toList();
+    else
+      for (var i = 0;
+          i < widget.drawers.length && i < menuControllers.length;
+          i++) {
+        menuControllers[i].direction = widget.drawers[i].direction;
+      }
+    for (var i = menuControllers.length; i < widget.drawers.length; i++) {
+      menuControllers.add(createController(widget.drawers[i]));
+    }
     if (widget.controller != null) {
       widget.controller._menuControllers = menuControllers;
       widget.controller._setFocus = (index) {
@@ -189,7 +198,6 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
   int drawerFrom(Direction direction) {
     return menuControllers.indexWhere((element) {
-      
       return element.direction == direction;
     });
   }
@@ -239,7 +247,6 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
               if (details.globalPosition.dx < maxSlideAmount + 60) {
                 int focusDrawer = drawerFrom(Direction.left);
 
-                
                 if (focusDrawer < 0) {
                 } else {
                   this.focusDrawerIndex = focusDrawer;
@@ -253,8 +260,6 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
                   details.globalPosition.dx > width - maxSlideAmount - 60) {
                 int focusDrawer = drawerFrom(Direction.right);
 
-                
-
                 if (focusDrawer < 0) {
                   return;
                 } else {
@@ -266,11 +271,9 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
                     startDx = details.globalPosition.dx;
                 }
               }
-              
             },
             onHorizontalDragUpdate: (details) {
               if (startDx == -1) return;
-              
 
               double dx = (details.globalPosition.dx - startDx);
               MenuController menuController = menuControllers[focusDrawerIndex];
@@ -318,7 +321,7 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
             : WillPopScope(
                 child: content,
                 onWillPop: () {
-                  return new Future(() {
+                  return Future(() {
                     if (isDrawerOpen()) {
                       menuControllers.forEach((element) {
                         element.close();
@@ -356,14 +359,19 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
     double slideAmount = maxSlideAmount * slidePercent;
     final contentScale = 1.0 - ((1.0 - drawer.percentage) * scalePercent);
-    final cornerRadius = widget.cornerRadius * menuController.percentOpen;
-    
+    final cornerRadius = (drawer.cornerRadius ?? widget.cornerRadius) *
+        menuController.percentOpen;
 
-    if (widget.drawers[focusDrawerIndex].direction == Direction.right)
-      slideAmount = -slideAmount + (maxSlideAmount * (1 - contentScale));
+    if (drawer.degree != null) {
+      slideAmount = slideAmount * (1 - (1 - contentScale) / 2);
+      if (widget.drawers[focusDrawerIndex].direction == Direction.right) {
+        slideAmount = -slideAmount;
+      }
+    } else if (widget.drawers[focusDrawerIndex].direction == Direction.right)
+      slideAmount = -slideAmount;
+
     double degreeAmount = (drawer.degree ?? 0) * slidePercent;
     degreeAmount = degreeAmount * pi / 180;
-    
 
     Matrix4 perspective;
     if (drawer.degree == null) {
@@ -385,25 +393,29 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
       // }
     }
 
+    double elevation = drawer.elevation * slidePercent;
     return new Transform(
       transform: perspective,
-      origin: drawer.direction == Direction.right
+      origin: drawer.degree != null
           ? Offset(MediaQuery.of(context).size.width / 2, 0.0)
-          : null,
+          : drawer.direction == Direction.right
+              ? Offset(MediaQuery.of(context).size.width, 0.0)
+              : null,
       alignment: Alignment.centerLeft,
-      child: new Container(
-        decoration: new BoxDecoration(
-          boxShadow: widget.contentShadow,
-        ),
-        child: new ClipRRect(
-            borderRadius: new BorderRadius.circular(cornerRadius),
-            child: content),
+      child: Card(
+        elevation: elevation,
+        clipBehavior: Clip.antiAliasWithSaveLayer,
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(cornerRadius)),
+        margin: EdgeInsets.symmetric(horizontal: elevation),
+        child: content,
       ),
     );
   }
 
   @override
   Widget build(BuildContext context) {
+    focusDrawerIndex = min(widget.drawers.length - 1, focusDrawerIndex);
     return Stack(
       children: [
         focusDrawerIndex != null
@@ -458,7 +470,6 @@ class DrawerScaffoldMenuControllerState
 
   MenuController getMenuController(BuildContext context,
       [Direction direction = Direction.left]) {
-    
     final scaffoldState =
         context.findAncestorStateOfType<_DrawerScaffoldState>();
     return scaffoldState.menuControllers.firstWhere(
@@ -503,7 +514,7 @@ class Screen {
 class MenuController extends ChangeNotifier {
   final TickerProvider vsync;
   final AnimationController _animationController;
-  final Direction direction;
+  Direction direction;
   MenuState state = MenuState.closed;
 
   MenuController(